requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.29.0
pydantic>=2.5.1
python-dotenv>=1.0.0
faker>=19.13.0
typing-extensions>=4.8.0
llama-cpp-python>=0.2.0
openai>=1.3.0
